LayoutTests/fast/events/domnodeinserted-entire-fragment-expected.txt

DOMNodeInserted events should be fired after all nodes in a DocumentFragment are inserted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS container.childNodes.length is 3
PASS container.childNodes.length is 3
PASS container.childNodes.length is 3
PASS successfullyParsed is true

TEST COMPLETE
